Moved to long and storing values in Dictionary

+++ AdventOfCode2019/Day9/Day9PreviousIntCodeTests.cs
@@ -19,7 +19,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithOnlyOpCode99()
         {
-            var valueAt0 = new IntCode(new int[]{99}, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[]{99}, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(99);
         }
@@ -27,7 +27,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithAdditionOpCodeStopAfterOneIteration()
         {
-            var valueAt0 = new IntCode(new int[] { 01, 0, 0, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 01, 0, 0, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(2);
         }
@@ -35,7 +35,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithInsufficientNumbersForFinalOpCode()
         {
-            var valueAt0 = new IntCode(new int[] { 01, 5, 0, 0, 99, 5 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 01, 5, 0, 0, 99, 5 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(6);
         }
@@ -43,7 +43,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithMultiplicationOpCodeStopAfterOneIteration()
         {
-            var valueAt0 = new IntCode(new int[] { 02, 1, 1, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 02, 1, 1, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(1);
         }
@@ -51,7 +51,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithMultipleDifferentOperations()
         {
-            var valueAt0 = new IntCode(new int[] { 01, 1, 1, 4, 99, 5, 6, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 01, 1, 1, 4, 99, 5, 6, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(30);
         }
@@ -62,7 +62,7 @@ namespace AdventOfCode2019.Day9
             var input = 42;
             _input.Setup(x => x.Get()).Returns(input);
 
-            var valueAt0 = new IntCode(new int[] { 03, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 03, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(input);
         }
@@ -73,7 +73,7 @@ namespace AdventOfCode2019.Day9
             var input = 42;
             _input.Setup(x => x.Get()).Returns(input);
 
-            var valueAt0 = new IntCode(new int[] { 03, 0, 1, 0, 6, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 03, 0, 1, 0, 6, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(input + 99);
         }
@@ -81,7 +81,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithOutputOpCodeStopAfterOneIteration()
         {
-            var valueAt0 = new IntCode(new int[] { 04, 2, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 04, 2, 99 }, _input.Object, _output.Object)[0];
 
             _output.Verify(_ => _.Push(99));
         }
@@ -89,7 +89,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void IntcodeWithOutputOpCodeDoesNotStop()
         {
-            var valueAt0 = new IntCode(new int[] { 04, 4, 1, 2, 6, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 04, 4, 1, 2, 6, 0, 99 }, _input.Object, _output.Object)[0];
 
             _output.Verify(_ => _.Push(6));
             valueAt0.ShouldBe(100);
@@ -98,7 +98,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void ParametersToAddForAddInstructionAreImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1101, 5, 5, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1101, 5, 5, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(10);
         }
@@ -106,7 +106,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void ParametersToMultiplyForMultiplyInstructionAreImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1102, 5, 5, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1102, 5, 5, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(25);
         }
@@ -114,7 +114,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OutputParameterImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 104, 25, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 104, 25, 99 }, _input.Object, _output.Object)[0];
 
             _output.Verify(_ => _.Push(25));
         }
@@ -122,7 +122,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void NegativeValuesAsImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1101, 100, -1, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1101, 100, -1, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(99);
         }
@@ -130,7 +130,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void ProgramConsideredFinishedIfHaltAfterOutput()
         {
-            var intCode = new IntCode(new int[] { 04, 0, 99 }, _input.Object, _output.Object);
+            var intCode = new IntCode(new long[] { 04, 0, 99 }, _input.Object, _output.Object);
 
             intCode.Finished.ShouldBeTrue();
         }
@@ -138,7 +138,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void ProgramNotConsideredFinishedIfHaltAfterAddition()
         {
-            var intCode = new IntCode(new int[] { 01, 0, 0, 0, 99 }, _input.Object, _output.Object);
+            var intCode = new IntCode(new long[] { 01, 0, 0, 0, 99 }, _input.Object, _output.Object);
 
             intCode.Finished.ShouldBeFalse();
         }
@@ -146,7 +146,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void ProgramNotConsideredFinishedIfHaltAfterMultiply()
         {
-            var intCode = new IntCode(new int[] { 02, 0, 0, 0, 99 }, _input.Object, _output.Object);
+            var intCode = new IntCode(new long[] { 02, 0, 0, 0, 99 }, _input.Object, _output.Object);
 
             intCode.Finished.ShouldBeFalse();
         }
@@ -154,7 +154,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfTrueWithFalseAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 05, 2, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 05, 2, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(5);
         }
@@ -162,7 +162,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfTrueWithTrueAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 05, 3, 4, 10, 5, 1101, 4, 4, 0, 99}, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 05, 3, 4, 10, 5, 1101, 4, 4, 0, 99}, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(8);
         }
@@ -170,7 +170,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfTrueWithFalseAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 105, 0, 1, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 105, 0, 1, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(105);
         }
@@ -178,7 +178,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfTrueWithTrueAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1105, 1, 3, 1101, 4, 3, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1105, 1, 3, 1101, 4, 3, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(7);
         }
@@ -186,7 +186,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfFalseWithFalseAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 06, 2, 1, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 06, 2, 1, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(6);
         }
@@ -194,7 +194,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfFalseWithTrueAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 06, 3, 9, 0, 1101, 53, 1, 0, 99, 4}, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 06, 3, 9, 0, 1101, 53, 1, 0, 99, 4}, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(54);
         }
@@ -202,7 +202,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfFalseWithFalseAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 106, 1, 1, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 106, 1, 1, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(106);
         }
@@ -210,7 +210,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeJumpIfFalseWithTrueAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1106, 0, 4, 99, 1101, 3, 8, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1106, 0, 4, 99, 1101, 3, 8, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(11);
         }
@@ -218,7 +218,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeLessThanWithFalseAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 07, 6, 5, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 07, 6, 5, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(0);
         }
@@ -226,7 +226,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeLessThanWithTrueAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 07, 5, 6, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 07, 5, 6, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(1);
         }
@@ -234,7 +234,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeLessThanWithFalseAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1107, 2, 1, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1107, 2, 1, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(0);
         }
@@ -242,7 +242,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeLessThanWithTrueAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1107, 1, 2, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1107, 1, 2, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(1);
         }
@@ -250,7 +250,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeEqualsWithFalseAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 08, 6, 5, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 08, 6, 5, 0, 99, 1, 2 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(0);
         }
@@ -258,7 +258,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeEqualsWithTrueAndPositional()
         {
-            var valueAt0 = new IntCode(new int[] { 08, 5, 6, 0, 99, 1, 1 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 08, 5, 6, 0, 99, 1, 1 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(1);
         }
@@ -266,7 +266,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeEqualsWithFalseAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1108, 2, 1, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1108, 2, 1, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(0);
         }
@@ -274,7 +274,7 @@ namespace AdventOfCode2019.Day9
         [Fact]
         public void OpCodeEqualsWithTrueAndImmediate()
         {
-            var valueAt0 = new IntCode(new int[] { 1108, 2, 2, 0, 99 }, _input.Object, _output.Object)[0];
+            var valueAt0 = new IntCode(new long[] { 1108, 2, 2, 0, 99 }, _input.Object, _output.Object)[0];
 
             valueAt0.ShouldBe(1);
         }
@@ -284,9 +284,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(8);
 
-            new IntCode(new int[] { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1));
         }
 
@@ -295,9 +295,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(7);
 
-            new IntCode(new int[] { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 9, 8, 9, 10, 9, 4, 9, 99, -1, 8 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(0));
         }
 
@@ -306,9 +306,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(8);
 
-            new IntCode(new int[] { 3, 3, 1107, -1, 8, 3, 4, 3, 99 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 3, 1107, -1, 8, 3, 4, 3, 99 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(0));
         }
 
@@ -317,9 +317,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(7);
 
-            new IntCode(new int[] { 3, 3, 1107, -1, 8, 3, 4, 3, 99 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 3, 1107, -1, 8, 3, 4, 3, 99 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1));
         }
 
@@ -328,9 +328,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(0);
 
-            new IntCode(new int[] { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(0));
         }
 
@@ -339,9 +339,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(10);
 
-            new IntCode(new int[] { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1));
         }
 
@@ -350,9 +350,9 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(10);
 
-            new IntCode(new int[] { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 }, _input.Object, _output.Object);
+            new IntCode(new long[] { 3, 3, 1105, -1, 9, 1101, 0, 0, 12, 4, 12, 99, 1 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1));
         }
 
@@ -361,11 +361,11 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(7);
 
-            new IntCode(new int[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
+            new IntCode(new long[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
                 1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,
                 999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(999));
         }
 
@@ -374,11 +374,11 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(8);
 
-            new IntCode(new int[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
+            new IntCode(new long[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
                 1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,
                 999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1000));
         }
 
@@ -387,11 +387,11 @@ namespace AdventOfCode2019.Day9
         {
             _input.Setup(_ => _.Get()).Returns(9);
 
-            new IntCode(new int[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
+            new IntCode(new long[] { 3,21,1008,21,8,20,1005,20,22,107,8,21,20,1006,20,31,
                 1106,0,36,98,0,0,1002,21,125,20,4,20,1105,1,46,104,
                 999,1105,1,46,1101,1000,1,20,4,20,1105,1,46,98,99 }, _input.Object, _output.Object);
 
-            _output.Verify(o => o.Push(It.IsAny<int>()), Times.Once);
+            _output.Verify(o => o.Push(It.IsAny<long>()), Times.Once);
             _output.Verify(o => o.Push(1001));
         }
     }
