+++ AdventOfCode2019/Day9/IOutput.cs
@@ -2,6 +2,6 @@ namespace AdventOfCode2019.Day9
 {
     public interface IOutput
     {
-        void Push(int outputValue);
+        void Push(long outputValue);
     }
 }
